requirements.txt
ansible
beautifulsoup4
boto3
chatgpt
colorama
configparser
countryinfo
discord.py
docker
dpkt
dronekit-sitl
dumper
ExifRead
faker
flask
folium
geocoder
geoip2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
hexy
huggingface_hub
importlib-metadata
isbntools
keyboard
langchain-community
langchain-core
langchain
langchain_openai
matplotlib
mobi
mysql-connector-python
networkx
opencv-python
openpyxl
osmnx
packaging
pandas
passlib
pdf2docx
pexpect
phonenumbers
pillow
platformdirs
plotly
pluggy
pokebase
praw
protobuf
psutil
pushbullet.py
py
pyautogui
pyOpenSSL
pyautogui
pycparser
pycryptodome
pydrive2
pygame
PyGithub
pyjokes
pylint
pymongo
pymysql
pyopenssl
pyparsing
pypdf
pyTelegramBotAPI
pytest
python-dateutil
python-docx
python-dotenv
python-magic
python-nmap
python-pptx
pyttsx3
PySocks
pywhatkit
pyyaml
rarfile
rfc3986
redis
requests
RedDownloader
seaborn
selenium
SimpleGUITK
six
slack-bolt
slack-sdk
sniffio
sortedcontainers
sounddevice
soundfile
speechrecognition
sqlalchemy
statsmodels
tk
tomli
tomlkit
trio
trio-websocket
typing_extensions
urllib3
virtualbox
wifi-qrcode-generator
wikipedia-api
wrapt
wsproto
xmltodict
yagmail
yamllint
yarl
zipfile36
